+++ src/classes/year.test.ts
@@ -12,13 +12,21 @@ import "../../__mocks__/hooks";
 import Year from "./year";
 import Month from "./month";
 import {Weekday} from "./weekday";
-import {GameSystems, GameWorldTimeIntegrations, LeapYearRules, TimeKeeperStatus, YearNamingRules} from "../constants";
+import {
+    GameSystems,
+    GameWorldTimeIntegrations,
+    LeapYearRules,
+    PredefinedCalendars,
+    TimeKeeperStatus,
+    YearNamingRules
+} from "../constants";
 import LeapYear from "./leap-year";
 import Season from "./season";
 import Moon from "./moon";
 import SimpleCalendar from "./simple-calendar";
 import {Note} from "./note";
 import {SimpleCalendarConfiguration} from "./simple-calendar-configuration";
+import PredefinedCalendar from "./predefined-calendar";
 
 describe('Year Class Tests', () => {
     let year: Year;
@@ -1076,6 +1084,10 @@ describe('Year Class Tests', () => {
         expect(data.name).toBe('Winter');
         data = year.getSeason(2, 9);
         expect(data.name).toBe('Summer');
+
+        year.seasons[0].startingMonth = -1;
+        data = year.getSeason(0, 1);
+        expect(data.name).toBe('Spring');
     });
 
     test('Get Year Name', () => {
@@ -1096,6 +1108,34 @@ describe('Year Class Tests', () => {
         year.yearNamingRule = YearNamingRules.Random;
         expect(year.getYearName(4)).not.toBe('');
 
+    });
+
+    test('Get Sunrise Sunset Time', () => {
+        year.months.push(month);
+        let sunrise = year.getSunriseSunsetTime(year.numericRepresentation, month, month.days[0]);
+        expect(sunrise).toBe(0);
+
+        PredefinedCalendar.setToPredefined(year, PredefinedCalendars.Gregorian);
+
+        SimpleCalendar.instance = new SimpleCalendar();
+        SimpleCalendar.instance.currentYear = year;
+
+        sunrise = year.getSunriseSunsetTime(year.numericRepresentation, year.months[0], year.months[0].days[0], true, false);
+        expect(sunrise).toBe(21600);
+
+        sunrise = year.getSunriseSunsetTime(year.numericRepresentation, year.months[2], year.months[2].days[19], false, false);
+        expect(sunrise).toBe(64800);
+
+        sunrise = year.getSunriseSunsetTime(year.numericRepresentation, year.months[3], year.months[3].days[19], true, true);
+        expect(sunrise).toBe(1618898400);
+
+        sunrise = year.getSunriseSunsetTime(year.numericRepresentation, year.months[11], year.months[11].days[21], true, false);
+        expect(sunrise).toBe(21600);
+
+        year.seasons[1].startingMonth = 3;
+        year.seasons[1].startingDay = 29;
+        sunrise = year.getSunriseSunsetTime(year.numericRepresentation, year.months[0], year.months[0].days[0], true, false);
+        expect(sunrise).toBe(21600);
     });
 
 });